fix: Use lazy type to specialize `Edge`/`Connection`
Use a lazy type to properly resolve and specialize the `Edge`/`Connection`
types that are created from the model.

The reason we are using a subclass of `LazyType` is because the original
one expects a path to be passed to be used to import it later. We don't
have one here, but we do have a mapper containing all the created types
that can be retrieved later.

Fix #97